+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 `psql` with the `-X` flag to prevent any `.psqlrc` commands from
 accidentally triggering the load of a previous DB version.**
 
+## Latest
+
+**Minor features**
+* #2736 Support adding columns to hypertables with compression enabled
+
 ## 2.0.0 (2020-12-18)
 
 With this release, we are officially moving TimescaleDB 2.0 to GA,